rsyslogd to log to /data/rsyslog
Summary:
This changeset updates the rsyslog configuration so that the `messages`
and `secure` logs are output to `/data/rsyslog`, instead of the default
`/var/log` as that is not persistent.  Related fixes for log rotation
and other rsyslogd rules are also included.

Fixes T329, T330, T335.

Test Plan: Manual on MT7623 systems.

Reviewers: dp, duvall

Maniphest Tasks: T335, T330, T329

Differential Revision: https://ph0.b10e.net/D631