Endpoint for verifying OAuth ID Token
Summary:
In my MacOS app, once you complete the OAuth flow, you are left with an ID token.
I created an endpoint that my app would post the ID token to, so that we can first
verify the token and then decode the contents of the token and send it back to the app.

Test Plan: Tested with Dan's scripts, queried the endpoints from my MacOS app.

Reviewers: dp, stephen

Reviewed By: stephen

Subscribers: buildbot

Differential Revision: https://ph0.b10e.net/D633