Require all config tree changes to be part of a rollback changeset
Summary:
All changes to the config tree should be performed within a Changeset
Init/Commit pair, even if there is no reason the change would fail or be
rolled back.  This is needed to maintain the integrity of the internal
tree structures.  The two places where this constraint was being
violated have been fixed.  The tree code will complain if it detects the
constraints being violated in the future.

Test Plan: manual

Differential Revision: https://ph0.b10e.net/D641