Don't request certs by fingerprint; just ask for a cert
Summary:
This removes the ability of an appliance to ask for an arbitrary certificate as
long as it knows the hash of that cert.

It also simplifies the retrieval by removing the handshake: the appliance now
simply polls the cloud for a new cert, rather than being told via the config
tree as soon as one is available.

Old clients will still ask for a cert by fingerprint, but will get back
whatever they get back.  They don't check that the fingerprint of the
received cert matches what they asked for, so that's safe.  For now, they will
continue to be notified of updated certificates, but this support (all the
changes in `cl-cert/cert.go`) should be pulled once we know all the clients are
past this point.

This also includes a fix for working out of a proto area, where an extra
`$APROOT` would be prepended to the certificate paths, and a fix to allow `mcp`
to put its logfile at an arbitrary path.

Fixes T342.

Test Plan:
Run an old client against the new cloud bits, and vice versa.  See that with new
client bits we get updated certificates at `ap.rpcd` start, or periodically
thereafter, and that with old client bits we get them immediately.

Reviewers: dp

Reviewed By: dp

Subscribers: buildbot

Maniphest Tasks: T342

Differential Revision: https://ph0.b10e.net/D639