T351 Improve robustness and diagnosibility
Summary:
When we kill a process for consuming too much memory, use SIGABRT to
dump the go stacktraces.  This may help understand how that daemon got
into that state.

When restarting a daemon, any dependent daemons that are marked as
'broken' will have that state cleared.  If the daemon's brokenness was a
side effect of the parent being broken, it's possible that restarting
the parent will also fix the child.

Test Plan:
manually breaking, killing, and restarting daemons.  Also
modified mcp.json to lower the memory limits for daemons to trigger the
SIGABRT kills.

Differential Revision: https://ph0.b10e.net/D650