Avoid races in appliance-side certificate acquisition
Summary:
When a box comes up for the first time, `ap.httpd`, `ap.userauthd`, and
`ap.rpcd` will race to produce a certificate; the first two will generate a
self-signed cert, and the last will pull one from the cloud.  `ap.httpd` starts
late, after the cloud cert is available, so it's fine, but `ap.userauthd` starts
before `ap.rpcd`, so the cloud certificate becomes available while it's
generating that self-signed cert, which it then proceeds to use, because it
hasn't set up the config handler yet to be notified of the change.  Meanwhile
the config tree certificate `state` nodes are incorrect.

This change causes `ap.httpd` and `ap.userauthd` to check for certs, but not to
generate them; that's left up to `ap.rpcd`, just like the fetching of the cloud
cert.  `ap.httpd` and `ap.userauthd` will wait if no certs are available, and be
restarted when `ap.rpcd` notifies them of a cert (regardless of origin).

Fixes T349

Test Plan:
Reproduce the bad behavior on an alpha box.  Put the rebuilt `ap.rpcd`,
`ap.httpd`, and `ap.userauthd` onto the box and re-run the test scenario.  See
that `ap.rpcd` doesn't finish generating the self-signed cert before it can
download the cloud cert, and appropriately notifies the other daemons, which
then restart and find the cloud cert.

Reviewers: dp

Reviewed By: dp

Subscribers: nils, buildbot

Maniphest Tasks: T349

Differential Revision: https://ph0.b10e.net/D645