Add some settings for the VSCode Go plugin
Summary:
This enables use of the language server (`gopls`), sets the default
`$GOPATH` to point to our `golang` folder, so that navigation works
propertly, and sets the formatting tool to `gofmt` to match what we use
(though use of `gopls` supplants all other tools for formatting, if
enabled).

Test Plan:
Run VSCode with the Go plugin enabled and hop from one symbol to its
definition, etc.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D663