Don't clutter the user's known_hosts file with tunnel keys
Summary:
As a person runs `cl-service` and connects to the appliance on the other end,
`$HOME/.ssh/known_hosts` ends up getting cluttered with the host ssh keys
generated by `ap.rpcd`.  Since they're one-time use keys, they'll keep getting
added to the file, never to be used again.

We can tell ssh to put the key in a different file, in the temporary directory
where we already put the private key and config file.

Test Plan:
Build and copy `cl-service` and `ssh_config.got` to `tun0` and run it out of my
home directory.  Verify that my personal `known_hosts` file hasn't been touched,
but the one in `/tmp/...` is.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D664