egress filtering: block spoofed source IPs
Summary:
Drop any packet from a client device that appears to have a spoofed source
IP address.  At the moment, this simply means that the source IP doesn't
belong to the originating subnet.  If required, we could ensure that the
source IP matches the source MAC as well, but that would require more
frequent updates to, and reloads of, the iptables config file.

Test Plan:
Used 'source NATting' iptables rules on a Linux client to
spoof its source address, and verified that those packets were blocked
by the firewall.

Differential Revision: https://ph0.b10e.net/D666