Put chrony's drift file in /data on mt7623
Summary:
On mt7623, chrony complains about being unable to write its driftfile.  This is
because it doesn't have sufficient permissions to write in the directory where
the file is configured to go.  Plus that directory is on tmpfs.

This makes the driftfile location a platform-specific piece of configuration,
puts it in `/data` on mt7623, and makes sure that the directory gets created.

Fixes T336

Test Plan:
Upgrade package on mt7623 and rpi.  On mt7623, see that the drift directory gets
created with correct permissions, that chrony's config file gets written to
point to it, and that the drift file eventually gets placed there.  On rpi, see
that the config file continues to point to the original place, and that the
existing drift file does get overwritten.

Reviewers: dp

Reviewed By: dp

Maniphest Tasks: T336

Differential Revision: https://ph0.b10e.net/D667