T267 default @/network/dnsserver is invalid
Summary:
Added a new "IP address with optional port" config type.  This correctly
validates the current dnssserver setting (8.8.8.8:53).  It allows users
to specify just an IP address for a service, implying that we should use
the default port for the service, or they can specify both an IP address
and a non-default port.

Test Plan:
Added new tests to configd_test.go to exercise both valid
and invalid settings to @/network/dnsserver.

Differential Revision: https://ph0.b10e.net/D683