enable root SSH password login for Alpha
Summary:
This changeset amends the packaging scripts for `bg-appliance` to relax
the `sshd`(8) configuration so that an operator with the `root` password
can connect.  We rely on the default firewall properties that further
restrict SSH to the `standard` and `core` trust groups.

Fixes T357.

Test Plan: Repeated `opkg` install and remove operations.

Reviewers: dp, duvall, nils

Maniphest Tasks: T357

Differential Revision: https://ph0.b10e.net/D677