Add throttled logging for some lively messages
Summary:
There are a handful of messages (generally error conditions) that may be
issued repeatedly and persistently, without adding any new information.
This adds a wrapper around the zap logger which will regulate the
frequency with which these redundant messages are logged.

Test Plan:
manual, but temporarily adding code to issue a message every
half second forever, and verified that it was throttled appropriately.
Tunneled into nils-wrt from tun0 to regression test the changes to the
tunnel logic.

Differential Revision: https://ph0.b10e.net/D687