T388 configd needs to load expirations at startup
Summary:
When ap.configd starts, or has its full tree replaced by the cloud, all
properties with pending expiration times need to be inserted into the
expiration heap.  This functionality seems to have been lost when moving
the tree construction code into a common module, so we are currently
only processing expirations that are created or updated while the daemon
is running.

The logic around resetting the expiration handler's clock has been
simplified.  The handler now has sole custody of that clock.  Other go
routines will use a new channel to signal that the clock should be
reevaluated.

Test Plan:
Manual.  Added multiple properties with different expiration times.
Verified that the expirations were correctly processed, both while the
daemon was running and when the daemon was restarted.  Verified that the
expiration mechanism was robust across multi-week time jumps.

Differential Revision: https://ph0.b10e.net/D688