Try to get a reasonable system time before appliance stack runs
Summary:
As described in T370, this change

  - causes `ap.mcp` to periodically touch a file in `/etc` so that on reboot,
    `sysfixtime` can reset the time to something reasonably new
  - configures `chronyd` statically, so that it can fetch the time prior to
    `ap.mcp` starts (`ap.networkd` will overwrite that configuration)
  - causes `ap.mcp` to wait for up to 30 seconds for the clock to sync before
    starting up, trying to prevent any time jump from happening while the
    appliance stack is running

Fixes T370, T389.

Test Plan:
Manual.  Upgrade an alpha box running bits prior to this to these bits and see
that the clock step happens in `chronyd`'s initial run, prior to our stack
coming up, and that all the new config files are written correctly, and that
when `chronyd` restarts, it sees all of those as expected.  Also boot the box
without a WAN link, to simulate being unable to connect to the external time
servers, as might be done via a firewall, and see that our stack does start up,
even if it does take a bit longer.  Run the bits on a Pi for sanity as well.

Reviewers: dp, stephen

Reviewed By: dp, stephen

Maniphest Tasks: T389, T370

Differential Revision: https://ph0.b10e.net/D689