T392 repeated log message took down ap.mcp and the appliance
Summary:
ap.mcp will now detect and throttle repeated log messages.  This should
reduce the risk of filling /data.

There are no messages that we expect ap.mcp to issue both repeatedly and
immediately, so any such occurrence suggests that something is broken
within the daemon.  To prevent this brokenness from burning all the CPU
(disrupting service and preventing debugging), we will use these
messages as an indication that we should throttle the daemon as well.

When configd replaces its entire config tree with a backup from the
cloud, we need all of the daemons to refresh any internal state based on
config data.  To accomplish this, configd will exit 30 seconds after
getting the new tree.  This will prompt ap.mcp to restart configd and
all daemons that depend on it.  The 30 second delay leaves time for
ap.rpcd to notify cl.rpcd that the REFRESH command was successfully
executed.

Test Plan:
added a tight loop in ap.mcp that issued the same messages
hundreds of times before moving to a new message.  Verified that the
message display and daemon were both throttled correctly.

Deleted ap_props and backup.  Verified that they were refreshed from the
cloud, and that the daemons restarted correctly.

Differential Revision: https://ph0.b10e.net/D694