ap-factory install to other partition, including packages
Summary:
This changeset is a significant enhancement to `ap-factory`, giving it
more awareness of which partition is inactive, the ability to mount or
unmount that partition in its overlay configuration, and to install to
that partition, including a list of packages.

Test Plan: Multiple retrievals and installs on MT7623 systems.

Reviewers: dp, duvall, nils

Differential Revision: https://ph0.b10e.net/D678