ap.watch sampler should try harder to open an interface
Summary:
Currently a vlan's ap.watch sampler thread will exit if the interface it
is meant to sample doesn't appear to be present.  Because bridge
interfaces are constructed dynamically by ap.networkd, and because
ap.networkd tells ap.mcp that it is online before those bridges are
fully constructed, ap.watchd may decide to leave some vlans unsampled.

Rather than giving up if an interface appears to be missing or offline,
the sampler threads will now continue to retry the open every second
until the interface is available or the daemon exists.

Unrelated small change: the map of throttled loggers in
aputil/logging.go needs to be protected by a mutex.

Test Plan:
manually restarting ap.networkd and ap.watchd multiple times
to verify that all interfaces are properly sampled.

Differential Revision: https://ph0.b10e.net/D700