Detect exim RCE vulnerability CVE-2019-10149
Summary:
All versions of the exim mail daemon between 4.87 and 4.91 are
vulnerable to a remote code execution attack.  We already detect an
earlier exim vulnerability.  This change builds on that mechanism to
detect the versions vulnerable to the latest CVE as well.

Test Plan:
Stood up a debian server running exim 4.89 and verified that
both the vulnerabilities were detected and reported.  Upgraded
to exim 4.92 and verified that the vulnerability config entry
was cleared.

Reviewers: cat, dp

Reviewed By: cat, dp

Subscribers: dp

Differential Revision: https://ph0.b10e.net/D701