Revised device listing page
Summary:
This revises the device listing page to be more information dense.  We include
icons to indicate if a client is wireless or wired, the IP address, and a red
"chip" if there are alerts, with a count of the number.

Stop showing inactive devices by default.  Add a checkbox to make them appear.

Fixes T248
Ref T233

Test Plan:
- Ran against mock and live data
- Ran for a few minutes as live web-app on svc1

Reviewers: stephen

Reviewed By: stephen

Maniphest Tasks: T248, T233

Differential Revision: https://ph0.b10e.net/D744