Add breadcrumbs indicating Organization and Site names, as a navigational aid
Summary:
We believe that as we transition to a UI which can support multiple organizations,
each with multiple sites, that users may become confused about where they are
in the app.  This adds contextual information to help the user know where they
are.

Rework each page's navbar title to indicate the purpose of the page;
this also means less chance of overflow.  Other context is left to the page
content area where overflow is better managed.

Test Plan: - Test in appliance and cloud mode.

Reviewers: stephen

Reviewed By: stephen

Subscribers: fiona

Tags: #doc

Differential Revision: https://ph0.b10e.net/D704