crash configd if it becomes non-responsive
Summary:
If a daemon is unable to communicate with configd, the daemon
essentially becomes useless.  If the problem is on the client side, the
repeated disconnect/reconnect attempts in apcfg.go:sendOp() should clear
it up.  If the problem is inside configd, then it may last until the box
is rebooted.

This adds a HealthMonitor goroutine to the apcfg module.  If the daemon
opts in, this routine will track communications errors and ask ap.mcp to
crash configd if enough errors are detected.  This crash will result in
configd's stacktraces to be logged, ideally aiding in postmortem
debugging.  In addition, all daemons the depend on configd will restart,
clearing any issues or stale data that resulted from configd being
unavailable for a period of time.

Test Plan:
added code to configd so it would stop responding to messages
after 30 seconds.  Verified that watchd was able to successfully trigger
the crash after reporting multiple errors.

Differential Revision: https://ph0.b10e.net/D711