Reduce size of appliance stats archives
Summary:
The size of the stats archives being uploaded to the cloud are sufficiently
large that post-processing tools are now running out of memory on smaller
VMs.  This includes a number of changes to reduce the size of data to
address that issue, reduce bandwith consumption from the appliance, and
storage consumption in the cloud.

  - Remove the 'port:service' map from the DeviceRecord structure.  This is
    a giant wad of text that was rolled into the struct as part of the
    defaultpass project.  The 'port' part of the map is included elsewhere
    in the struct, the 'service' text isn't used in the cloud, and it can
    be regenerated anyway.

  - Add support for the binary 'gob' format in addition to the ASCII json
    format.

  - Add a custom gob marshaler for the Drop data, which compresses 17-byte
    ascii ip:port strings into the 6-byte binary equivalent.

In addition to the data reduction, this adds the appliance UUID field to
each line of the exported CSV data, allowing us to filter statistics from a
merged database, based on the site.

Test Plan:
This has been running in the cloud every night for a month or
so.  It's currently running on nils-plateau.

Differential Revision: https://ph0.b10e.net/D709