Add per-client rolling averages for network activity
Summary:

Take frequent snapshots of the per-client network usage statistics, and use
them to calculate rolling averages of traffic per-day, hour, minute, and
second.

Store these averages in a new @/metrics branch of the config tree.  Because
of the frequency of updates to this branch, it is not persisted and changes
do not trigger Update events.  Because there are no update events, a
cloud-based client must poll the appliance for changes.

Test Plan: manual on nils-wrt in the office and svc0 in the cloud

Differential Revision: https://ph0.b10e.net/D707