Allow clients to migrate to lower-level trust groups
Summary:
When a client connects to a VAP which doesn't host its
assigned ring, that client should be assigned to the default ring for
that VAP.  This relaxes the current policy, which is to automatically
assign clients to higher-level groups, but to require an administrative
action to migrate to lower-level groups.  This new policy is less secure
than the original, but it is intended to help reduce some of the pain
being felt at early alpha customers.

Tangentially related: the first time a client is assigned a DHCP address
in a new ring, the lease is only valid for one minute.  On renewal, it
will be granted a full-duration lease.  This helps mitigate a common
race condition, where a client's DHCP request packet is issued before
hostapd is finished migrating it to the correct VLAN, and before its
final ring assigment has been determined.

Test Plan:
manual testing on nils-wrt with a mac client, migrating
between all three SSIDs.

Differential Revision: https://ph0.b10e.net/D712