support bind mount for /data
Summary:
This changeset adds a bind mount of `/data` to the alternate root constructed by `ap-factory`, such
that package operations on `/data` (such as the daemon directory tree) are
performed against the actual shared partition.  This second mount means
that content delivered in `/data` must be safe for multiple versions of
the software accessing it.

Test Plan: Build; testing on MT7623 in progress.

Reviewers: dp, duvall, nils

Reviewed By: duvall, nils

Tags: #sw-dev

Differential Revision: https://ph0.b10e.net/D715