Make sure to create the full cert directory structure
Summary:
Creating the certificate-specific subdir of `$APSECRET/ssl` fails if the parent
`ssl` directory isn't there, so be sure to use `os.MkdirAll()`.

Test Plan:
Remove `/data/secret/ssl` from the filesystem and `@/certs` from the config tree
on a test box, and confirm that existing code ends up generating a self-signed
cert, while being repeatedly unable to create the directory for the cloud cert.
With the new code, the directory creation succeeds.

Reviewers: dp, stephen

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D718