T391 Allow a ring's programmatic subnet to be overridden manually
Summary:
Ring subnets are generated programatically, based on the site index and
base_address properties.  This adds a per-ring config property, allowing
that programatic subnet to be overridden.  This is intended to be a
rarely used feature to simplify the integration of legacy devices.  As
such, it should require brightgate support to enable it, and should not
be surface in the UI.

Test Plan: manual with both wired and wireless devices

Differential Revision: https://ph0.b10e.net/D726