Make sure that the validation tree is itself valid
Summary:
It is possible to add properties to the config tree where the type of a
node label (not the value of a leaf node) can be confused with the label
of one or more of its siblings.  This problem should be detected
immediately, rather than sporadically during the use of that part of the
config tree.

Fixes T419

Test Plan: Automated tests pass.

Reviewers: nils, dp

Reviewed By: nils

Maniphest Tasks: T419

Differential Revision: https://ph0.b10e.net/D731