Don't leak gRPC stream resources when failing to fetch a stream
Summary: Fixes T421

Test Plan:
Build `cl.rpcd` to always fail in `FetchStream()` and run the rebuilt `ap.rpcd`
against it.  See that it doesn't grow in RSS, and that the number of goroutines
doesn't increase (after instrumenting it to log `runtime.NumGoroutine()`),
unlike without this fix.

Reviewers: dp

Reviewed By: dp

Maniphest Tasks: T421

Differential Revision: https://ph0.b10e.net/D734