Generate self-signed certs before cloud; renew self-signed certs
Summary:
This lets cloudless setups work, since we don't have to wait for cloud
connectivity before generating a self-signed cert.  We also regenerate the cert
once it expires, assuming a cloud cert hasn't replaced it.  And we extend the
duration of the self-signed certs from a week to a month.

We also no longer stop the generation part-way, in favor of a more robust system
of dealing with the race against cloud cert retrieval.

In addition, there are miscellaneous fixes to deal with truncated self-signed
certs and private keys as well as cleanup of code that never really worked
properly.

Fixes T409, T410

Test Plan:
Manual on x86.  Slow the cert generation down so that it can meaningfully race
against the cert retrieval and see that the generated cert loses correctly.
Truncate the self-signed key and cert so that they're properly regenerated.  Let
the cloud cert expire and see that a new self-signed cert is installed, and see
that a new cloud cert will replace the self-signed cert before the latter's
expiration.

Reviewers: dp

Reviewed By: dp

Subscribers: nils

Maniphest Tasks: T410, T409

Differential Revision: https://ph0.b10e.net/D732