T297 Apply simple fault patterns to MT7623 LED3
Summary:
Adds a new daemon, ap.tron, which monitors some fundamental system
status and uses two LEDs on the box to communicate that status to the
user.  This status is intended to help the user/installer to diagnose
failures that prevent the system from phoning home (this preventing
tunneling) and/or from running the local web management interface.

Led 3 is used to convey network status:
    - Have an IP address
    - Have network connectivity to the world
    - DNS name resolution works

Led 4 conveys brightgate stack status:
    - ap.tron is runnning
    - ap.mcp is running and responsive
    - ap.configd is running and responsive
    - ap.rpcd is communicating with the cloud

The status elements reported on Led 3 are also reflected in the
@/metrics/health config subtree (when configd is running), allowing them
to be displayed by the management UI.

Test Plan: manual on nils-wrt and a satellite

Differential Revision: https://ph0.b10e.net/D749