Fix segfault when cert state node is missing expiration
Summary:
This fixes a bug we ran into at one customer.  The bug (how the expiration got
set to nil in the first place) has not been root caused, but I've added some
logging for when we get into this situation, in the hopes that it'll help with
the diagnosis.  The expiration will be re-added when it's found to be missing.

Test Plan:
On a cloud-simulated appliance, get a new cert, remove the expiration both
before and after it's been installed, and see that `ap.rpcd` doesn't crash, but
does properly log the condition.

Reviewers: nils, dp

Reviewed By: nils, dp

Differential Revision: https://ph0.b10e.net/D759