Replace ZeroMQ with a Go-native Nanomsg implementation
Summary:
ZeroMQ is a C++ communications library with numerous idiosyncracies and
workarounds in our code, and which doesn't seem to work correctly on the
next OpenWRT build.  This replaces ZMQ with Mangos, which is a pure Go
implementation of the Nanomsg socket library.

The new library provides many of the same abstractions, but in a more
straightforward and robust way.  For example, if an error in a
request/reply exchange results in the reply being lost, the ZMQ
connection would be left in an inconsistent and unrecoverable state,
which would require tearing down and reestablishing the connection,
which could require restarting all of our daemons.

Test Plan:
initial system-level testing on nils-wrt.  Currently running
on dogfood and satellites.

Differential Revision: https://ph0.b10e.net/D763