ap-factory shouldn't create `dpath` in current directory
Summary:
When `ap-factory install` tests to see if `/data` is what it expects it to be,
it tries to ensure the directories `/data/configd` and `/data/secret/rpcd` are
present, and will create them if not.

Unfortunately, it tests for the presence of, and then creates, `dpath`, in the
current directory; the variable containing the path inadvertently had quotes put
around it.

Test Plan:
Run `ap-factory install -n ...`, as the directories are made even in dry-run
mode.

Reviewers: dp, stephen

Reviewed By: stephen

Differential Revision: https://ph0.b10e.net/D762