Add scans for conficker-related vulnerabilities
Summary:
This adds two new nmap invocations to our vulnerability list:

    CVE-2008-4250: checks whether the device is vulnerable to the
    conficker worm.

    conficker: checks whether the device has been infected by the
    conficker worm.

This also splits the default password scan out from the rest of the
vulnerability scans.  The password scan takes significantly longer than
the others, and needs to be run much less frequently.

Removed the nmap script that probed for ms17-010, as it is bundled with
the new version of nmap.

Test Plan:
Created a new Windows XP VM which does not have the patch to
fix CVE-2008-4250.  Verified that the vulnerability was detected.  I
couldn't find a sample of the conficker worm anywhere, so I haven't been
able to test a successful detection.  I have verified that the scan
doesn't report a false positive for a clean XP image.

Differential Revision: https://ph0.b10e.net/D768