Add support for inter-ring Airplay to AppleTV
Summary:
Adds a new rule that allows the wide range of UDP traffic that Airplay
rquires.  In the long run, this really needs to be restricted just to
AppleTV devices.

Allow rules to filter on both source and destination ports.

Cleans up the rules that we were passing to iptables.  The old rules
were correct, but more wordy than necessary.

Test Plan:
manual testing to verify that Airplay works and that legacy
rules still behave as expected.  Added automated regression testing for
the firewall rule parser and generator.

Differential Revision: https://ph0.b10e.net/D771