Minimal functionality for ap.rpcd even without cloud access
Summary:
If ap.rpcd can't reach the cloud or find its cloud credentials, it will
come up in FAILSAFE mode.  This will allow it to create local certs,
allowing ap.httpd and ap.userauthd to function.

ap.httpd and ap.userauthd will both enter FAILSAFE mode while they are
sleeping, waiting for certs to be made available.  When the certs are
created, they will wake up and enter ONLINE mode.

Test Plan:
manual testing on nils-wrt, with and without a cloud secret,
and with/without cloud and local certs already on the system.

Differential Revision: https://ph0.b10e.net/D773