Collect fault information and push to the cloud
Summary:
This introduces a framework for generating and storing a variety of
fault information on the appliance and periodically uploading it to the
cloud.  This information should help debug appliances that are known to
be misbehaving.  It will also help us identify failures/crashes that are
happening on the appliance, but which are going unreported because
daemons are automatically restarted.

Each fault is stored as a json document in single file in /data/faults
on the appliance.  Each dameon may generate faults for itself, and mcp
may generate faults for any daemon.

Periodically ap.rpcd will scan the faults directory for new fault
reports, pushing them to the cloud.  Faults will be retained on the
appliance for several days to facilitate local debugging.  After that
point, reports that have been pushed to the cloud will be deleted from
the appliance.

cl.eventd will receive the fault reports and will store them in
per-appliance directories.  Still to do: persist the reports to cloud
storage.

This initial commit includes three types of faults:

    - crashes: mcp detects an unexpected daemon exit.  The fault report
      includes the last 32k of logs from that daemon, which should include
      any stack trace along with recent activity

    - memory: mcp detects that a daemon is using excessive memory.

    - errors: catch-all category for failures detected by daemons which
      are logged, should be investigated, but don't require the daemon
      to crash.  These are generally "should not happen" events, which
      may indicate software errors.  This initial commit only includes a
      small sampling of the many places where these fault events could
      be invoked,

Test Plan: manual testing with nils-wrt and svc0.b10e.net

Differential Revision: https://ph0.b10e.net/D778