Appliance upgrade mechanism
Summary:
This provides a minimal implementation of a release mechanism for the Brightgate
fleet.  Most actions are manual: definitions of releases, decisions about what
releases to upgrade what appliances to, and when to make that happen.  But once
told (through the config tree) that an upgrade is available, an appliance will
download the appropriate artifacts, checksum them, install them, and reboot.

Test Plan:
Automated tests for the database code; manual testing for the rest against both
a BG-100 (as a gateway) and a Pi (as the BG-100's satellite).

Reviewers: dp, stephen

Reviewed By: dp, stephen

Differential Revision: https://ph0.b10e.net/D765