introduce classification properties
Summary:
This changeset introduces classification properties to the appliance
configuration tree, in advance of the delivery of the classifier
infrastructure.  Each classifier rated as "production ready" (as opposed
to still in "experiment") has an associated string-valued classification
property.  These properties are managed by a configuration writer that
is aware of the classifier's associated probability with the
classification and only changes (or clears) the classification value
when the probability is above (or below) an associated threshold.

The properties for the previous classifier, `@/clients/[mac]/identity` and
`@/clients/[mac]/confidence`, will be removed in a future changeset.

The three properties, placed under `@/clients/[mac]/classification`, are
- `interface_manufacturer`, set by a "lookup" classifier based on the IEEE OUI database,
- `device_genus`, set by a Bayesian classifier trained on Brightgate data, and
- `os_genus`, also set by a Bayesian classifier trained on Brightgate data.

Test Plan:
Built and tested on MT7623.  Manual internal `add`, `set`, `del`
operations on new properties.

Reviewers: dp, duvall, nils

Differential Revision: https://ph0.b10e.net/D786