simple utility to test internet speed
Summary:
This change introduces a utility, `ap-speedtest`, and an appliance endpoint,
`/check/speedtest`, which allow a logged-in user or someone connected to the
appliance, respectively, to check the speed of the appliance's connection to
the internet.

Test Plan:
Run `ap-speedtest` manually on x86 and mt7623, see that the numbers are more or
less as expected for those machines' connections to the internet.  Point a
browser at `ap.httpd` on x86 and see that `/check/speedtest` returns the same
data.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D781