Roles presentation for the frontend doesn't work right
Summary:
The roles presentation in the frontend isn't right for users outside of
the logged-in user's home organization.  Switch to presenting role
limit-sets entirely per-account.  This requires a rather substantial
rework of the database queries and /api/account/:acct_uuid/roles REST
endpoint.  Even though this is broken in the frontend, the backend
refuses to grant such roles (thanks to the database design).

Test Plan:
- Unit testing
- Frontend testing using simulated cloud
- Frontend testing using mock data

Reviewers: stephen, duvall

Reviewed By: stephen, duvall

Differential Revision: https://ph0.b10e.net/D782