Put wireless neighborhood reporting into /check/diag
Summary:
We want the results of a wireless scan in the diagnostics output so we can see
what the neighborhood looks like -- how crowded it is, what channels are being
used, what features nearby APs have, etc.

Test Plan:
Access `/check/diag` with `curl` on the gateway's command-line (requires putting
`ap.httpd` into developer mode) and with a browser connected wirelessly to the
gateway.  Only for mt7623.

Reviewers: dp, stephen

Reviewed By: dp, stephen

Differential Revision: https://ph0.b10e.net/D790