deauth devices and rebuild user list on user deletion
Summary:
When deleting a user, we would like to deauthenticate all devices that
are actively using that user's identity to connect to our network.
Because we do not currently have a user->device mapping, we
deauthenticate all devices connected to an EAP VAP.  We also reset the
RADIUS hostapd instance, forcing it to reconstruct its users file.

Test Plan: manual testing on nils-wrt

Differential Revision: https://ph0.b10e.net/D792