Allow ap-rpc to configure itself from configd
Summary:
When running `ap-rpc`, you either need to want to connect to the stock
`cl.rpcd`, or you need to configure it on the commandline.  Since it's invoked
in `ap.httpd`, via `/check/diag`, that's not always possible.

This change makes it check `@/cloud/svc_rpc/0/...` for its connectivity
settings, just like the daemon does, except that if it can't connect to
`ap.configd` (or its prerequisite, `ap.brokerd`), it continues on with the
command-line or built-in configurations.

Fixes T457

Test Plan: Manual testing on x86.

Reviewers: dp, nils

Reviewed By: dp, nils

Maniphest Tasks: T457

Differential Revision: https://ph0.b10e.net/D791